backend/requirements.txt
Flask
Flask-CORS
pdfplumber
python-docx
